move utils function in internal folder
