+++ v-next/hardhat-chai-matchers/src/internal/matchers/emit.ts
@@ -1,4 +1,4 @@
-import type { AssertWithSsfi, Ssfi } from "../../utils.js";
+import type { AssertWithSsfi, Ssfi } from "../utils/ssfi.js";
 import type { EventFragment } from "ethers/abi";
 import type { Contract } from "ethers/contract";
 import type { Provider, TransactionReceipt } from "ethers/providers";
@@ -9,9 +9,9 @@ import util from "node:util";
 import { HardhatError } from "@ignored/hardhat-vnext-errors";
 import { AssertionError } from "chai";
 
-import { buildAssert } from "../../utils.js";
 import { ASSERTION_ABORTED, EMIT_MATCHER } from "../constants.js";
 import { assertArgsArraysEqual, assertIsNotNull } from "../utils/asserts.js";
+import { buildAssert } from "../utils/build-assert.js";
 import { preventAsyncMatcherChaining } from "../utils/prevent-chaining.js";
 
 export const EMIT_CALLED = "emitAssertionCalled";
