+++ v-next/hardhat-chai-matchers/src/internal/matchers/reverted/reverted.ts
@@ -3,7 +3,7 @@ import type { HardhatEthers } from "@ignored/hardhat-vnext-ethers/types";
 import { HardhatError } from "@ignored/hardhat-vnext-errors";
 import { numberToHexString } from "@ignored/hardhat-vnext-utils/hex";
 
-import { buildAssert } from "../../../utils.js";
+import { buildAssert } from "../../utils/build-assert.js";
 import { REVERTED_MATCHER } from "../../constants.js";
 import { assertIsNotNull } from "../../utils/asserts.js";
 import { preventAsyncMatcherChaining } from "../../utils/prevent-chaining.js";
