+++ v-next/hardhat-chai-matchers/src/internal/matchers/reverted/revertedWith.ts
@@ -1,7 +1,7 @@
 import { HardhatError } from "@ignored/hardhat-vnext-errors";
 import { numberToHexString } from "@ignored/hardhat-vnext-utils/hex";
 
-import { buildAssert } from "../../../utils.js";
+import { buildAssert } from "../../utils/build-assert.js";
 import { REVERTED_WITH_MATCHER } from "../../constants.js";
 import { preventAsyncMatcherChaining } from "../../utils/prevent-chaining.js";
