+++ v-next/hardhat-chai-matchers/src/internal/matchers/reverted/revertedWithCustomError.ts
@@ -1,16 +1,16 @@
-import type { Ssfi } from "../../../utils.js";
+import type { Ssfi } from "../../utils/ssfi.js";
 import type { ErrorFragment, Interface } from "ethers/abi";
 import type { BaseContract } from "ethers/contract";
 
 import { HardhatError } from "@ignored/hardhat-vnext-errors";
 import { numberToHexString } from "@ignored/hardhat-vnext-utils/hex";
 
-import { buildAssert } from "../../../utils.js";
 import {
   ASSERTION_ABORTED,
   REVERTED_WITH_CUSTOM_ERROR_MATCHER,
 } from "../../constants.js";
 import { assertArgsArraysEqual, assertIsNotNull } from "../../utils/asserts.js";
+import { buildAssert } from "../../utils/build-assert.js";
 import { preventAsyncMatcherChaining } from "../../utils/prevent-chaining.js";
 
 import {
