+++ v-next/hardhat-chai-matchers/src/internal/matchers/reverted/revertedWithPanic.ts
@@ -2,7 +2,7 @@ import { HardhatError } from "@ignored/hardhat-vnext-errors";
 import { toBigInt } from "@ignored/hardhat-vnext-utils/bigint";
 import { numberToHexString } from "@ignored/hardhat-vnext-utils/hex";
 
-import { buildAssert } from "../../../utils.js";
+import { buildAssert } from "../../utils/build-assert.js";
 import { REVERTED_WITH_PANIC_MATCHER } from "../../constants.js";
 import { preventAsyncMatcherChaining } from "../../utils/prevent-chaining.js";
